helper: Visit directory entries in order.
Previously in febootstrap-supermin-helper we would visit the
files in supermin.d in arbitrary (ie. readdir) order.  This has
caused a series of heisenbugs where some implicit dependency
between these files has not been honoured.  The latest one is
that '/etc/localtime' can be added to the appliance before '/etc'
has been created (and this operation fails).

Instead of continuing to chase these, this commit forces us to
visit the files in filename order by sorting them before visiting
them.

Note that in libguestfs, the current order is sufficient, because
the files are called:

base.img
daemon.img
hostfiles