De-duplicate directories with the same name, like "/etc" and "/etc/"
In Rawhide there are two patches which create /etc/pam.d and /etc/pam.d/

febootstrap was trying to treat these as two separate directories.
Add a de-duplication step so that these names are considered the same.

We could go further than this patch (eg. removing multiple //
characters), but this patch is sufficient to fix Fedora builds.