Don't pass use_installed to every package handler function.
use_installed is a global variable (defined in febootstrap_cmdline.mli)
so there's not much point in passing it around to every function that
needs it.  This commit removes the optional argument in favour of just
using the global variable in each package handler.

However we still need a place where we can bail if the --use-installed
flag is used for package handlers which don't support this yet.

Thus add a ph_init function is called after the right package handler
has been detected but before it is used.  This is a convenient place
to put the --use-installed checking and any other initialization that
is required.