helper: Fix -u and -g options when host uses LDAP or NIS (RHBZ#835130).
The getpwnam/getgrnam interfaces are not very well specified in the
case where we are looking up a name that does not exist.

Previously the code assumed that if errno was set, that would
distinguish an error from the not found case.  However this is
certainly not true for LDAP, where errno == ENOENT might indicate not
found.  Other errno values could be returned too.

The specific problem is that if the user specifies a numeric -u or -g
option, then the first lookup (eg) getpwnam ("42") could return NULL
with a non-zero errno, and that would not indicate an error.

Change the parsing of both -u and -g options as follows:

(1) If getpwnam/getgrnam returns an error, record it in saved_errno,
but do NOT fail at this point (as we did before).

(2) If the name is numeric, return the numeric value regardless of
whether getpwnam/getgrnam returned any error.

(3) Otherwise print an error message and fail.  Ensure that
saved_errno is printed for debugging purposes.