Add febootstrap-supermin-helper --copy-kernel option.
Because libvirt wants to set an SELinux label on the kernel, we can't
just symlink to it (because we don't want to change the SELinux label
on the kernel in /boot, and in any case it's owned by root so we
cannot).

Add the --copy-kernel option so that libguestfs can indicate that it
wants to copy the kernel instead of symlinking it.