always work with empty --root in zypp driver
The current mode of operation in the zypp driver is to download all
packages into a temporary pkg-cache-dir. Installed packages in the root
filesystem are ignored, otherwise zypper would download only a subset of
packages, and it would not provide a list of all required dependencies.

Remove the getuid check to force an empty root directory and use the
configured system repositories.

This works for root and non-root users.

Signed-off-by: Olaf Hering <olaf@aepfle.de>