debian: Fix !use_installed case (RHBZ#970689).
Not sure if this code ever worked, but there were multiple
problems if the name of the package supplied on the command
line was not a simple filename (eg. contained '/').

When unpacking into the temporary directory, use the package
basename, not the full path name, as a component of the
temporary directory.

Thanks: Gabriel de Perthuis