helper: Fix confusing verbose message.
Previously the messages would say:

checking modpath /lib/modules/3.9.9-302.fc19.armv7hl.lpae is a directory
picked vmlinuz-3.9.9-302.fc19.armv7hl.lpae because modpath /lib/modules/3.9.9-302.fc19.armv7hl.lpae exists

However this appears to indicate that the kernel was "picked".  In
fact, it just passed the modpath test, and it's not necessarily the
kernel that we're going to pick.