helper: Implement --dtb / device tree search.
For architectures such as ARM which require a device tree, this
implements a supermin-helper --dtb option which can be used to search
for a device tree file matching a wildcard.

Usage example:

  supermin-helper -f ext2 --dtb 'vexpress-*a9' supermin-d -o /tmp

would find a suitable dtb and symlink it as /tmp/dtb (or exit with an
error is none could be found).