rpm: Flush debugging messages in C code.
Because C and OCaml use different implementations of stdio, and
because C stdout is not line-buffered, we can get mixed up debugging
messages.

Partially fix this by using fflush (stdout) after printing C debugging
messages (this doesn't completely fix it because we'd have to do the
same in OCaml code too).

This slows down debugging output, but it's only used when debugging so
we don't care about that.