rpm: Choose providers better (RHBZ#1266918).
In the referenced bug, a customer had installed a web browser called
'palemoon'.  The RPM of this web browser provides and requires various
core libraries, such as:

  Provides: libnss3.so()(64bit) # normally provided by 'nss'
  Requires: libxul.so()(64bit)  # normally provided by 'firefox'

Our previous algorithm -- inherited from the days when we used to run
'rpm' commands -- takes every provider of a particular requirement and
adds it to a big list, so if any other package requires
'libnss3.so()(64bit)', then both 'nss' and 'palemoon' are added to the
package list.

Yum used to handle this differently - it used to only pick the package
with the shortest name.  Later on, before yum was retired, it had a
more complex decision algorithm described here:
http://yum.baseurl.org/wiki/CompareProviders

This change makes supermin use the shortest name algorithm, so in the
case above, it always picks 'nss' over 'palemoon'.

There is a second possible problem which is not fixed by the current
patch set: If a package both provides and requires the same
dependency, we should ignore that dependency.  For example, 'palemoon'
both Provides and Requires 'libxul.so()(64bit)', so if 'palemoon' is
pulled in, then 'firefox' would not be an additional requirement.
Because we store all the packages in a big list, we lose track of
where a dependency originates from, so it is not easy to implement
this second change.