init: Uncompress modules before adding them to the mini initrd.
When building the mini initrd, previously we copied the modules into
the initrd as-is, so for example if the module was xz-compressed, we
copied the foo.ko.xz file to the initrd.  This requires that the mini
init binary is linked to zlib & lzma, so that it knows how to
uncompress these modules when insmoding them at boot time.  Also since
the init is statically linked, it required _static_ versions of these
libraries.

This changes things so that the modules are uncompressed in the mini
initrd, so they are a little bit larger, but the init binary no longer
needs to be statically linked to zlib & lzma.

The init binary is smaller (966K -> 837K), but because we are storing
uncompressed modules in the mini initrd, the initrd as a whole becomes
larger (1.4M -> 2.6M)

However there are benefits to this change:

 - The code in the init binary is much simpler.

 - Removes the dependency on static zlib & lzma.

 - We can use an alternate libc to make a much smaller init binary
   (see following commits).