docs: Document how to link the init binary with an alternate libc.
Because of the previous commits, it is now possible to link the init
binary using alternate libc's.  This updates the documentation to
describe how to link it using dietlibc or musl-libc.

Using dietlibc reduces the overall size of the initrd substantially.
For libguestfs, 2.6M -> 1.8M.  Though this is still slightly larger
than the initrd using compressed modules (1.4M).