init: Drop mbcache.ko.
This module is described as "Filesystem Meta Information Block Cache
(mbcache).  The mbcache caches blocks of block devices that need to be
located by their device/block number, as well as by other criteria
(such as the block's contents)."

It is compiled into the kernel in Fedora, and if it is required on
other distros then it should be pulled in by module dependencies.