ext2: Don't load whole files into memory when copying to the appliance (RHBZ#1113065).
Obviously for very large files this is going to be a problem, as well
as not being very cache efficient.

libext2fs can handle writes to parts of files just fine so copy files
in blocks.

Also demote the "Permission denied" error to a warning, and add some
explanatory text telling people not to use sudo.