Remove the --dtb option and all supporting code.
It used to be necessary to specify a DTB manually back in very old
versions of qemu on armv7.

However qemu has constructed a correct DTB for a very long time and
it's almost certainly a mistake if you are using the supermin --dtb
parameter to override that.  (For this reason, supermin will
intentionally give an error now if you try to use --dtb).

libguestfs stopped using this parameter in
commit af762e43424382d895d3a10aa6632adc601fdbad (2015-03-05).