init: Require root=... parameter on kernel command line.
In old versions of supermin, we picked the last device as the root
device.  Since 2012, libguestfs has passed the correct device to use
in the root= parameter, and this was supported by supermin (actually,
febootstrap!) >= 3.15.

This change requires the root= parameter to be passed, and supermin
will fail with an error otherwise.