kernel: Replace how kernel version information is found.
The old code was fairly bogus.  In the new code there is one "public"
function:

  val get_kernel_version : bool -> string -> string option

which takes a full vmlinuz path and returns the kernel version from it
if possible (or None if not).  This function first looks at the file
content to see if the version can be extracted, and if that fails
looks at the filename.

As a side effect of making this change, we also filter out kernels
from /boot where we cannot read the kernel version.