kernel: Don't allow End_of_file exception to escape if kernel is bogus (RHBZ#1477758).
If the kernel file is truncated (zero length or too short) then
functions such as ‘really_input’ raise End_of_file.  Unfortunately
this exception could escape and kill the program.  What we want to do
instead is to ignore such kernels.  The combination of this commit and
the last commit ignores such kernels if they are found in /boot.

Reported by Charlie Martinez.