Switch binary embedding to a C source
The current approach of writing a custom assembler snippet has few
issues:
- the assembler needs to manually things like the correct section,
  disabling the executable stack, and so on
- the symbols are global, and exported with an awkward interface
  (i.e. uint8_t offsets)
- the assembler source is built with a manual rule
- the embedded data is not properly aligned according to the needs of
  the platform
- in the past, the snippet had also tweaks for specific architectures
  (e.g. HPPA)

To overcome this situation, turn the bin2s.pl script into a bin2c.pl,
which instead generates a C snippet containing a C array with the data,
and a variable with the length of the data.  The code produced this way
has various advantages:
- it is simple C code, with no need for architecture tweaks
- it is constant, so the compiler places it in the rodata section, with
  the correct alignment
- it is static, so no symbols are exported
- it is included directly into the source using it, so no need for
  manual compilation rules

This is inspired by what the xxd tool (part of vim) does, although it is
less flexible than needed.