build: Pass CFLAGS & LDFLAGS to final supermin link (RHBZ#1624175).
We also use -runtime-variant _pic which selects the OCaml runtime
linked with -fPIC.  This will cause a performance regression on i686
although that probably doesn't matter now.

A bigger issue is that it will stop supermin from building with older
versions of OCaml (<= 4.02.2).  We might instead try detecting if it's
the old version in ./configure but that gets a bit fragile.