prepare: do not create base.tar.gz with no config files
tar defaults to 10K as block size, and thus it pads the created archives
to multiples of that size using zero's.  Hence, when creating an archive
with no files, it will be 10K zero's, that is compressed by gzip,
resulting in few bytes.  The issue happens later, during the build
phase: base.tar.gz is correctly detected as gz, and zcat is run to
detect its content: since the empty tar was 10K of zero's, the buffer in
get_compressed_file_content will be filled by zero's, and
get_file_content will fail to detect anything.

As solution, at least for our own base.tar.gz: in case there are no
config files to copy, do not create base.tar.gz at all.