rpm: fix version comparison
When comparing two RPMs, also the epoch, and the release must be taken
into account.  Hence, add a new helper to get the EVR string of a
package, and use it when sorting each list of installed packages.

The mapping is done to avoid recreating the EVR strings at every
comparison.