build: check for outputs in --if-newer check (RHBZ#1813809)
In case we need to check whether the appliance must be rebuilt, check
also the timestamp of the outputs of the mode, i.e. the kernel, initrd,
and root files. This way, when either of these files does not exist or
is older than the package manager DB we can rebuild the appliance.

Add a simple test to verify this behaviour.