src: Fix --if-newer --copy-kernel
We previously copied the kernel into the appliance using 'cp -p' which
preserves the datestamps of the installed kernel.  This can confuse
the --if-newer calculation, if for example the package database is
newer than the date on the installed kernel (which quite often is the
case).  This makes it think that the appliance is always older than
the package database, thus forcing a rebuild.

We can fix this using 'cp' instead of 'cp -p'.  We don't need the
permissions and datestamps on the copied kernel to be preserved anyway
(in fact, it could cause problems if the permissions are restrictive).

Fixes: commit 30de2cb603cdde33524a66d5466f6a9b986ce8a6