ocamlc: Use -output-complete-exe instead of -custom
This prevents bytecode executables from being broken by strip and
similar tools.  Note this is incompatible with OCaml < 4.10 (so breaks
RHEL 8).  However this only affects bytecode builds which we prefer
not to use in RHEL.  I left the old option in the Makefile so that it
could be uncommented by someone using older OCaml + bytecode.  We need
this for OCaml 5.0 since that drops native backends (temporarily) for
riscv64, s390x and ppc64le.